rodio

Module decoder

Decodes samples from an audio file.

Modules

symphonia

Symphonia decoders types

Structs

Decoder

Source of audio samples from decoding a file.

LoopedDecoder

Source of audio samples from decoding a file that never ends. When the end of the file is reached the decoder starts again from the beginning.

Enums

DecoderError

Error that can happen when creating a decoder.

Mp4Type