rodio

Module microphone

Microphone input support for audio recording. A microphone Source

Basic Usage

use rodio::microphone::MicrophoneBuilder;
use rodio::Source;
use std::time::Duration;

let mic = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .open_stream()?;

// Record audio for 3 seconds
let recording = mic.take_duration(Duration::from_secs(3));

// You can now use this like any other Source
// For example, play it back or process it further

Use preferred parameters if supported

Attempt to set a specific channel count, sample rate and buffer size but fall back to the default if the device does not support these

use rodio::microphone::MicrophoneBuilder;
use rodio::Source;
use std::time::Duration;

let mut builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .prefer_channel_counts([
        1.try_into().expect("not zero"),
        2.try_into().expect("not zero"),
    ])
    .prefer_sample_rates([
        16_000.try_into().expect("not zero"),
        32_000.try_into().expect("not zero"),
    ])
    .prefer_buffer_sizes(512..);

let mic = builder.open_stream()?;

Configuration with Error Handling

Attempt to set a specific channel count but fall back to the default if the device doesn’t support it:

use rodio::microphone::MicrophoneBuilder;
use rodio::Source;
use std::time::Duration;

let mut builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?;

// Try to set stereo recording (2 channels), but continue with default if unsupported
if let Ok(configured_builder) = builder.try_channels(2.try_into()?) {
    builder = configured_builder;
} else {
    println!("Stereo recording not supported, using default channel configuration");
    // builder remains unchanged with default configuration
}

let mic = builder.open_stream()?;

Device Selection

use rodio::microphone::{MicrophoneBuilder, available_inputs};

// List all available input devices
let inputs = available_inputs()?;
for (i, input) in inputs.iter().enumerate() {
    println!("Input {}: {}", i, input);
}

// Use a specific device (e.g., the second one)
let mic = MicrophoneBuilder::new()
    .device(inputs[1].clone())?
    .default_config()?
    .open_stream()?;

Structs

Input

An input device

InputConfig

Describes the input stream’s configuration

ListError

Error that can occur when we can not list the input devices

Microphone

A microphone input stream that can be used as Source.

MicrophoneBuilder

Builder for configuring and opening microphone input streams.

Enums

OpenError

Errors that can occur when opening a microphone source.

Functions

available_inputs

Returns a list of available input devices on the system.