rodio::microphone

Struct MicrophoneBuilder

pub struct MicrophoneBuilder<Device, Config, E = fn(StreamError)>
where
    E: FnMut(StreamError) + Send + Clone + 'static,
{ /* private fields */ }

Builder for configuring and opening microphone input streams.

Implementations

impl MicrophoneBuilder<DeviceNotSet, ConfigNotSet, fn(StreamError)>

pub fn new() -> MicrophoneBuilder<DeviceNotSet, ConfigNotSet, fn(StreamError)>

Creates a new microphone builder.

Example

let builder = rodio::microphone::MicrophoneBuilder::new();

impl<Device, Config, E> MicrophoneBuilder<Device, Config, E>
where E: FnMut(StreamError) + Send + Clone + 'static,

pub fn device( &self, device: Input, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigNotSet, E>, Error>

Sets the input device to use.

Example

let input = available_inputs()?.remove(2);
let builder = MicrophoneBuilder::new().device(input)?;

pub fn default_device( &self, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigNotSet, E>, Error>

Uses the system’s default input device.

Example

let builder = MicrophoneBuilder::new().default_device()?;

impl<Config, E> MicrophoneBuilder<DeviceIsSet, Config, E>
where E: FnMut(StreamError) + Send + Clone + 'static,

pub fn default_config( &self, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>, Error>

Uses the device’s default input configuration.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?;

pub fn config( &self, config: InputConfig, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>, Error>

Sets a custom input configuration.

Example

let config = InputConfig {
    sample_rate: NonZero::new(44_100).expect("44100 is not zero"),
    channel_count: NonZero::new(2).expect("2 is not zero"),
    buffer_size: cpal::BufferSize::Fixed(42_000),
    sample_format: cpal::SampleFormat::U16,
};
let builder = MicrophoneBuilder::new()
    .default_device()?
    .config(config)?;

impl<E> MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>
where E: FnMut(StreamError) + Send + Clone + 'static,

pub fn try_sample_rate( &self, sample_rate: SampleRate, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>, Error>

Sets the sample rate for input.

Error

Returns an error if the requested sample rate combined with the other parameters can not be supported.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .try_sample_rate(44_100.try_into()?)?;

pub fn prefer_sample_rates( &self, sample_rates: impl IntoIterator<Item = SampleRate>, ) -> MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>

Try multiple sample rates, fall back to the default it non match. The sample rates are in order of preference. If the first can be supported the second will never be tried.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    // 16k or its double with can trivially be resampled to 16k
    .prefer_sample_rates([
        16_000.try_into().expect("not zero"),
        32_000.try_into().expect("not_zero"),
    ]);

pub fn try_channels( &self, channel_count: ChannelCount, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>, Error>

Sets the number of input channels.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .try_channels(2.try_into()?)?;

pub fn prefer_channel_counts( &self, channel_counts: impl IntoIterator<Item = ChannelCount>, ) -> MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>

Try multiple channel counts, fall back to the default it non match. The channel counts are in order of preference. If the first can be supported the second will never be tried.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    // We want mono, if thats not possible give
    // us the lowest channel count
    .prefer_channel_counts([
        1.try_into().expect("not zero"),
        2.try_into().expect("not_zero"),
        3.try_into().expect("not_zero"),
    ]);

pub fn try_buffer_size( &self, buffer_size: u32, ) -> Result<MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>, Error>

Sets the buffer size for the input.

This has no impact on latency, though a too small buffer can lead to audio artifacts if your program can not get samples out of the buffer before they get overridden again.

Normally the default input config will have this set up correctly.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .try_buffer_size(4096)?;

pub fn prefer_buffer_sizes( &self, buffer_sizes: impl IntoIterator<Item = u32>, ) -> MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>

See the docs of try_buffer_size for more.

Try multiple buffer sizes, fall back to the default it non match. The buffer sizes are in order of preference. If the first can be supported the second will never be tried.

Note

We will not try buffer sizes larger then 100_000 to prevent this from hanging too long on open ranges.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .prefer_buffer_sizes([
        2048.try_into().expect("not zero"),
        4096.try_into().expect("not_zero"),
    ]);

Get the smallest buffer size larger then 512.

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .prefer_buffer_sizes(4096..);

impl<Device, E> MicrophoneBuilder<Device, ConfigIsSet, E>
where E: FnMut(StreamError) + Send + Clone + 'static,

pub fn get_config(&self) -> InputConfig

Returns the current input configuration.

Example

let builder = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?;
let config = builder.get_config();
println!("Sample rate: {}", config.sample_rate.get());
println!("Channel count: {}", config.channel_count.get());

impl<E> MicrophoneBuilder<DeviceIsSet, ConfigIsSet, E>
where E: FnMut(StreamError) + Send + Clone + 'static,

pub fn open_stream(&self) -> Result<Microphone, OpenError>

Opens the microphone input stream.

Example

let mic = MicrophoneBuilder::new()
    .default_device()?
    .default_config()?
    .open_stream()?;
let recording = mic.take_duration(Duration::from_secs(3)).record();