Function rodio::source::from_iter

pub fn from_iter<I>(iterator: I) -> FromIter<I::IntoIter> where
    I: IntoIterator, 

Builds a source that chains sources provided by an iterator.

The iterator parameter is an iterator that produces a source. The source is then played. Whenever the source ends, the iterator is used again in order to produce the source that is played next.

If the iterator produces None, then the sound ends.