Struct rodio::Sink

pub struct Sink { /* private fields */ }

Handle to a device that outputs sounds.

Dropping the Sink stops all sounds. You can use detach if you want the sounds to continue playing.

Implementations

impl Sink

pub fn try_new(stream: &OutputStreamHandle) -> Result<Sink, PlayError>

Builds a new Sink, beginning playback on a stream.

pub fn new_idle() -> (Sink, SourcesQueueOutput<f32>)

Builds a new Sink.

pub fn append<S>(&self, source: S)
where S: Source + Send + 'static, f32: FromSample<S::Item>, S::Item: Sample + Send,

Appends a sound to the queue of sounds to play.

pub fn volume(&self) -> f32

Gets the volume of the sound.

The value 1.0 is the “normal” volume (unfiltered input). Any value other than 1.0 will multiply each sample by this value.

pub fn set_volume(&self, value: f32)

Changes the volume of the sound.

The value 1.0 is the “normal” volume (unfiltered input). Any value other than 1.0 will multiply each sample by this value.

pub fn speed(&self) -> f32

Gets the speed of the sound.

The value 1.0 is the “normal” speed (unfiltered input). Any value other than 1.0 will change the play speed of the sound.

pub fn set_speed(&self, value: f32)

Changes the speed of the sound.

The value 1.0 is the “normal” speed (unfiltered input). Any value other than 1.0 will change the play speed of the sound.

pub fn play(&self)

Resumes playback of a paused sink.

No effect if not paused.

pub fn pause(&self)

Pauses playback of this sink.

No effect if already paused.

A paused sink can be resumed with play().

pub fn is_paused(&self) -> bool

Gets if a sink is paused

Sinks can be paused and resumed using pause() and play(). This returns true if the sink is paused.

pub fn clear(&self)

Removes all currently loaded Sources from the Sink, and pauses it.

See pause() for information about pausing a Sink.

pub fn skip_one(&self)

Skips to the next Source in the Sink

If there are more Sources appended to the Sink at the time, it will play the next one. Otherwise, the Sink will finish as if it had finished playing a Source all the way through.

pub fn stop(&self)

Stops the sink by emptying the queue.

pub fn detach(self)

Destroys the sink without stopping the sounds that are still playing.

pub fn sleep_until_end(&self)

Sleeps the current thread until the sound ends.

pub fn empty(&self) -> bool

Returns true if this sink has no more sounds to play.

pub fn len(&self) -> usize

Returns the number of sounds currently in the queue.