Module rodio::source

Sources of sound and various filters.

Structs

Amplify

Filter that modifies each sample by a given value.

BltFilter

Buffered

Iterator that at the same time extracts data from the iterator and stores it in a buffer.

ChannelVolume

Combines channels in input into a single mono source, then plays that mono sound to each channel at the volume given for that channel.

Delay

A source that delays the given source by a certain amount.

Done

When the inner source is empty this decrements an AtomicUsize.

Empty

An empty source.

EmptyCallback

An empty source which executes a callback function

FadeIn

Filter that modifies raises the volume from silence over a time period.

FromFactoryIter

Internal type used by from_factory.

FromIter

A source that chains sources provided by an iterator.

Mix

Filter that modifies each sample by a given value.

Pausable

PeriodicAccess

Calls a function on a source every time a period elapsed.

Repeat

A source that repeats the given source.

SamplesConverter

An iterator that reads from a Source and converts the samples to a specific rate and channels count.

SineWave

An infinite source that produces a sine.

SkipDuration

A source that skips specified duration of the given source from it’s current position.

Skippable

Spatial

Combines channels in input into a single mono source, then plays that mono sound to each channel at the volume given for that channel.

Speed

Filter that modifies each sample by a given value.

Stoppable

TakeDuration

A source that truncates the given source to a certain duration.

UniformSourceIterator

An iterator that reads from a Source and converts the samples to a specific rate and channels count.

Zero

An infinite source that produces zero.

Enums

SeekError

Traits

Source

A source of samples.

Functions

from_factory

Builds a source that chains sources built from a factory.

from_iter

Builds a source that chains sources provided by an iterator.

Type Aliases

Crossfade