Module rodio::source

Structs

Amplify
Filter that modifies each sample by a given value.

Buffered
Iterator that at the same time extracts data from the iterator and stores it in a buffer.

Delay
A source that delays the given source by a certain amount.

FadeIn
Filter that modifies each sample by a given value.

Mix
Filter that modifies each sample by a given value.

Repeat
A source that repeats the given source.

SineWave
An infinite source that produces a sine.

Speed
Filter that modifies each sample by a given value.

TakeDuration
A source that repeats the given source.

UniformSourceIterator
An iterator that reads from a Source and converts the samples to a specific rate and channels count.

Traits

Source
A source of samples.