rodio::source

Function from_factory

pub fn from_factory<F, S>(factory: F) -> FromIter<FromFactoryIter<F>> 
where
    F: FnMut() -> Option<S>,

Builds a source that chains sources built from a factory.

The factory parameter is a function that produces a source. The source is then played. Whenever the source ends, factory is called again in order to produce the source that is played next.

If the factory closure returns None, then the sound ends.