rodio

Struct Sink

pub struct Sink { /* private fields */ }

Handle to a device that outputs sounds.

Dropping the Sink stops all sounds. You can use detach if you want the sounds to continue playing.

Implementations

impl Sink

pub fn try_new(stream: &OutputStreamHandle) -> Result<Sink, PlayError>

Builds a new Sink, beginning playback on a stream.

pub fn new_idle() -> (Sink, SourcesQueueOutput<f32>)

Builds a new Sink.

pub fn append<S>(&self, source: S)
where S: Source + Send + 'static, f32: FromSample<S::Item>, S::Item: Sample + Send,

Appends a sound to the queue of sounds to play.

pub fn volume(&self) -> f32

Gets the volume of the sound.

The value 1.0 is the “normal” volume (unfiltered input). Any value other than 1.0 will multiply each sample by this value.

pub fn set_volume(&self, value: f32)

Changes the volume of the sound.

The value 1.0 is the “normal” volume (unfiltered input). Any value other than 1.0 will multiply each sample by this value.

pub fn speed(&self) -> f32

Changes the play speed of the sound. Does not adjust the samples, only the playback speed.

Note:

Increasing the speed will increase the pitch by the same factor

If you set the speed to 0.5 this will halve the frequency of the sound lowering its pitch.

If you set the speed to 2 the frequency will double raising the pitch of the sound.

Change in the speed affect the total duration inversely

If you set the speed to 0.5, the total duration will be twice as long.

If you set the speed to 2 the total duration will be halve of what it was.

See [Speed] for details

pub fn set_speed(&self, value: f32)

Changes the speed of the sound.

The value 1.0 is the “normal” speed (unfiltered input). Any value other than 1.0 will change the play speed of the sound.

Note:

Increasing the speed would also increase the pitch by the same factor

If you increased set the speed to 0.5, the frequency would be slower (0.5x the original frequency) .

Also if you set the speed to 1.5 the frequency would be faster ( 1.5x the original frequency).

Change in the speed would affect your total duration inversely

if you set the speed by 0.5, your total duration would be (2x the original total duration) longer.

Also if you set the speed to 2 the total duration would be (0.5 the original total_duration) shorter

pub fn play(&self)

Resumes playback of a paused sink.

No effect if not paused.

pub fn try_seek(&self, pos: Duration) -> Result<(), SeekError>

Attempts to seek to a given position in the current source.

This blocks between 0 and ~5 milliseconds.

As long as the duration of the source is known, seek is guaranteed to saturate at the end of the source. For example given a source that reports a total duration of 42 seconds calling try_seek() with 60 seconds as argument will seek to 42 seconds.

Errors

This function will return SeekError::NotSupported if one of the underlying sources does not support seeking.

It will return an error if an implementation ran into one during the seek.

When seeking beyond the end of a source this function might return an error if the duration of the source is not known.

pub fn pause(&self)

Pauses playback of this sink.

No effect if already paused.

A paused sink can be resumed with play().

pub fn is_paused(&self) -> bool

Gets if a sink is paused

Sinks can be paused and resumed using pause() and play(). This returns true if the sink is paused.

pub fn clear(&self)

Removes all currently loaded Sources from the Sink, and pauses it.

See pause() for information about pausing a Sink.

pub fn skip_one(&self)

Skips to the next Source in the Sink

If there are more Sources appended to the Sink at the time, it will play the next one. Otherwise, the Sink will finish as if it had finished playing a Source all the way through.

pub fn stop(&self)

Stops the sink by emptying the queue.

pub fn detach(self)

Destroys the sink without stopping the sounds that are still playing.

pub fn sleep_until_end(&self)

Sleeps the current thread until the sound ends.

pub fn empty(&self) -> bool

Returns true if this sink has no more sounds to play.

pub fn len(&self) -> usize

Returns the number of sounds currently in the queue.

pub fn get_pos(&self) -> Duration

Returns the position of the sound that’s being played.

This takes into account any speedup or delay applied.

Example: if you apply a speedup of 2 to an mp3 decoder source and get_pos() returns 5s then the position in the mp3 recording is 10s from its start.