Struct rodio::Sink

pub struct Sink {
    // some fields omitted
}

Handle to an endpoint that outputs sounds.

Dropping the Sink stops all sounds. You can use detach if you want the sounds to continue playing.

Methods

impl Sink

fn new(endpoint: &Endpoint) -> Sink

Builds a new Sink.

fn append<S>(&self, source: S) where S: Source + Send + 'static, S::Item: Sample, S::Item: Send

Appends a sound to the queue of sounds to play.

fn set_volume(&mut self, value: f32)

Changes the volume of the sound.

The value 1.0 is the "normal" volume (unfiltered input). Any value other than 1.0 will multiply each sample by this value.

fn detach(self)

Destroys the sink without stopping the sounds that are still playing.

fn get_min_remaining_duration(&self) -> Duration

Returns the minimum duration before the end of the sounds submitted to this sink.

Note that this is a minimum value, and the sound can last longer.

fn sleep_until_end(&self)

Sleeps the current thread until the sound ends.